rename and delete file
that's why edit-point-view looks like new one (previouse it was called create-point-view). In this commit I just update the part with eventHandlers
