+++ src/view/edit-point-view.js
@@ -1,14 +1,77 @@
+import { getDestinationById, getOffersVariantsByType } from '../mock/events';
+import { capitalize, transformToKebabCase } from '../utils';
 import AbstractView from '../framework/view/abstract-view';
 
-function createEditPointTemplate() {
+const BLANK_EVENT = {
+  type: '',
+  destination: '',
+  start: '',
+  end: '',
+  price: 0,
+  offers: []
+};
+
+function createDestinationOptionsTemplate(destinationList) {
+  let destinationOptions = '';
+  // здесь нельзя использовать const
+  // eslint-disable-next-line prefer-const
+  for (let destination of destinationList) {
+    destinationOptions += `<option value="${destination.name}"></option>`;
+  }
+
+  return `<datalist id="destination-list-1">${destinationOptions}</datalist>`;
+}
+
+function createOffersTemplate(type, offerIds) {
+  const allOffers = getOffersVariantsByType(type);
+  let offersContent = '';
+  // здесь нельзя использовать const
+  // eslint-disable-next-line prefer-const
+  for (let offer of allOffers) {
+    const isChecked = offerIds.includes(offer.id);
+    const offerTitleForAttr = transformToKebabCase(offer.title);
+
+    offersContent += `
+    <div class="event__offer-selector">
+      <input class="event__offer-checkbox  visually-hidden" id="event-${offerTitleForAttr}-1" type="checkbox" name="event-${offerTitleForAttr}" ${isChecked ? 'checked' : '' }>
+      <label class="event__offer-label" for="event-${offerTitleForAttr}-1">
+        <span class="event__offer-title">${offer.title} </span>
+        &plus;&euro;&nbsp;
+        <span class="event__offer-price"> ${offer.price}</span>
+      </label>
+    </div>`;
+  }
+
+  return `<div class="event__available-offers">${offersContent}</div>`;
+}
+
+function createPhotosTemplate(photos) {
+  let photosTemplate = '';
+  // здесь нельзя использовать const
+  // eslint-disable-next-line prefer-const
+  for (let photo of photos) {
+    photosTemplate += `<img class="event__photo" src="${photo.src}" alt="${photo.description}">`;
+  }
+
+  return `<div class="event__photos-container"><div class="event__photos-tape">${photosTemplate}</div></div>`;
+}
+
+
+function editPointFormTemplate(event, destinationVariants) {
+  const {type, dateFrom, dateTo, destination: destinationId, basePrice, offers: offerIds} = event;
+
+  const destination = getDestinationById(destinationId);
+  const detinationOptions = createDestinationOptionsTemplate(destinationVariants);
+  const offersTemplate = createOffersTemplate(type, offerIds);
+
   return `
   <li class="trip-events__item">
-  <form class="event event--edit" action="#" method="post">
+    <form class="event event--edit" action="#" method="post">
     <header class="event__header">
       <div class="event__type-wrapper">
         <label class="event__type  event__type-btn" for="event-type-toggle-1">
           <span class="visually-hidden">Choose event type</span>
-          <img class="event__type-icon" width="17" height="17" src="img/icons/flight.png" alt="Event type icon">
+          <img class="event__type-icon" width="17" height="17" src="img/icons/${type ? type : 'flight'}.png" alt="Event type icon">
         </label>
         <input class="event__type-toggle  visually-hidden" id="event-type-toggle-1" type="checkbox">
 
@@ -53,8 +116,8 @@ function createEditPointTemplate() {
             </div>
 
             <div class="event__type-item">
-              <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio"
-                name="event-type" value="check-in">
+              <input id="event-type-check-in-1" class="event__type-input  visually-hidden" type="radio" name="event-type"
+                value="check-in">
               <label class="event__type-label  event__type-label--check-in" for="event-type-check-in-1">Check-in</label>
             </div>
 
@@ -77,25 +140,21 @@ function createEditPointTemplate() {
 
       <div class="event__field-group  event__field-group--destination">
         <label class="event__label  event__type-output" for="event-destination-1">
-          Flight
+          ${capitalize(type)}
         </label>
         <input class="event__input  event__input--destination" id="event-destination-1" type="text"
-          name="event-destination" value="Chamonix" list="destination-list-1">
-        <datalist id="destination-list-1">
-          <option value="Amsterdam"></option>
-          <option value="Geneva"></option>
-          <option value="Chamonix"></option>
-        </datalist>
+          name="event-destination" value="${destination.name ? destination.name : ''}" list="destination-list-1">
+        ${detinationOptions}
       </div>
 
       <div class="event__field-group  event__field-group--time">
         <label class="visually-hidden" for="event-start-time-1">From</label>
         <input class="event__input  event__input--time" id="event-start-time-1" type="text" name="event-start-time"
-          value="18/03/19 12:25">
+          value="${dateFrom ? dateFrom : ''}">
         &mdash;
         <label class="visually-hidden" for="event-end-time-1">To</label>
         <input class="event__input  event__input--time" id="event-end-time-1" type="text" name="event-end-time"
-          value="18/03/19 13:35">
+          value="${dateTo ? dateTo : ''}">
       </div>
 
       <div class="event__field-group  event__field-group--price">
@@ -103,7 +162,7 @@ function createEditPointTemplate() {
           <span class="visually-hidden">Price</span>
           &euro;
         </label>
-        <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="160">
+        <input class="event__input  event__input--price" id="event-price-1" type="text" name="event-price" value="${basePrice ? basePrice : ''}">
       </div>
 
       <button class="event__save-btn  btn  btn--blue" type="submit">Save</button>
@@ -116,64 +175,14 @@ function createEditPointTemplate() {
       <section class="event__section  event__section--offers">
         <h3 class="event__section-title  event__section-title--offers">Offers</h3>
 
-        <div class="event__available-offers">
-          <div class="event__offer-selector">
-            <input class="event__offer-checkbox  visually-hidden" id="event-offer-luggage-1" type="checkbox"
-              name="event-offer-luggage" checked>
-            <label class="event__offer-label" for="event-offer-luggage-1">
-              <span class="event__offer-title">Add luggage</span>
-              &plus;&euro;&nbsp;
-              <span class="event__offer-price">50</span>
-            </label>
-          </div>
-
-          <div class="event__offer-selector">
-            <input class="event__offer-checkbox  visually-hidden" id="event-offer-comfort-1" type="checkbox"
-              name="event-offer-comfort" checked>
-            <label class="event__offer-label" for="event-offer-comfort-1">
-              <span class="event__offer-title">Switch to comfort</span>
-              &plus;&euro;&nbsp;
-              <span class="event__offer-price">80</span>
-            </label>
-          </div>
-
-          <div class="event__offer-selector">
-            <input class="event__offer-checkbox  visually-hidden" id="event-offer-meal-1" type="checkbox"
-              name="event-offer-meal">
-            <label class="event__offer-label" for="event-offer-meal-1">
-              <span class="event__offer-title">Add meal</span>
-              &plus;&euro;&nbsp;
-              <span class="event__offer-price">15</span>
-            </label>
-          </div>
-
-          <div class="event__offer-selector">
-            <input class="event__offer-checkbox  visually-hidden" id="event-offer-seats-1" type="checkbox"
-              name="event-offer-seats">
-            <label class="event__offer-label" for="event-offer-seats-1">
-              <span class="event__offer-title">Choose seats</span>
-              &plus;&euro;&nbsp;
-              <span class="event__offer-price">5</span>
-            </label>
-          </div>
-
-          <div class="event__offer-selector">
-            <input class="event__offer-checkbox  visually-hidden" id="event-offer-train-1" type="checkbox"
-              name="event-offer-train">
-            <label class="event__offer-label" for="event-offer-train-1">
-              <span class="event__offer-title">Travel by train</span>
-              &plus;&euro;&nbsp;
-              <span class="event__offer-price">40</span>
-            </label>
-          </div>
-        </div>
+        ${offersTemplate}
       </section>
 
       <section class="event__section  event__section--destination">
         <h3 class="event__section-title  event__section-title--destination">Destination</h3>
-        <p class="event__destination-description">Chamonix-Mont-Blanc (usually shortened to Chamonix) is a resort area
-          near the junction of France, Switzerland and Italy. At the base of Mont Blanc, the highest summit in the Alps,
-          it's renowned for its skiing.</p>
+        <p class="event__destination-description">${destination.description ? destination.description : ''}</p>
+
+        ${destination.pictures.length > 0 ? createPhotosTemplate(destination.pictures) : ''}
       </section>
     </section>
   </form>
@@ -182,7 +191,34 @@ function createEditPointTemplate() {
 }
 
 export default class EditPointView extends AbstractView {
-  get template() {
-    return createEditPointTemplate();
+  #event = null;
+  #destinationVariants = null;
+  #handleFormSubmit = null;
+  #handleCloseForm = null;
+
+
+  constructor({event = BLANK_EVENT, destinations = [], onFormSubmit, onCloseForm}) {
+    super();
+    this.#event = event;
+    this.#destinationVariants = destinations;
+    this.#handleFormSubmit = onFormSubmit;
+    this.#handleCloseForm = onCloseForm;
+
+    this.element.querySelector('.event--edit').addEventListener('submit', this.#formSubmitHandler);
+    this.element.querySelector('.event__rollup-btn').addEventListener('click', this.#formCloseHandler);
   }
+
+  get template() {
+    return editPointFormTemplate(this.#event, this.#destinationVariants);
+  }
+
+  #formSubmitHandler = (event) => {
+    event.preventDefault();
+    this.#handleFormSubmit();
+  };
+
+  #formCloseHandler = (event) => {
+    event.preventDefault();
+    this.#handleCloseForm();
+  };
 }
